Add Hooks namespace to maintain consistency with helpers
