+++ packages/victory-core/src/index.js
@@ -44,7 +44,7 @@ export * as Wrapper from "./victory-util/wrapper";
 export * as Axis from "./victory-util/axis";
 export { default as TimerContext } from "./victory-util/timer-context";
 export { default as PortalContext } from "./victory-portal/portal-context";
-export * from "./victory-util/hooks";
+export * as Hooks from "./victory-util/hooks";
 export * as LineHelpers from "./victory-util/line-helpers";
 // TODO: Fix this default export
 export { default as PointPathHelpers } from "./victory-util/point-path-helpers";
