+++ packages/victory-core/src/victory-util/hooks/index.js
@@ -1,0 +1,2 @@
+export { usePreviousProps } from "./use-previous-props";
+export { useAnimationState } from "./use-animation-state";
