+++ packages/victory-core/src/victory-util/hooks/use-previous-props.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const usePreviousProps = (props) => {
+  const ref = React.useRef();
+  React.useEffect(() => {
+    ref.current = props;
+  });
+  return ref.current || {};
+};
